Make StackNavigator parent
