+++ App.tsx
@@ -1,15 +1,12 @@
 import 'react-native-gesture-handler';
 import {GestureHandlerRootView} from 'react-native-gesture-handler';
 import React from 'react';
-import {NavigationContainer} from '@react-navigation/native';
-import TabNavigator from './src/navigation/TabNavigator.tsx';
+import StackNavigator from './src/navigation/StackNavigator.tsx';
 
 function App(): React.JSX.Element {
   return (
     <GestureHandlerRootView style={{flex: 1}}>
-      <NavigationContainer>
-        <TabNavigator />
-      </NavigationContainer>
+      <StackNavigator />
     </GestureHandlerRootView>
   );
 }
