+++ src/navigation/TabNavigator.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import CatalogStackNavigator from './CatalogStackNavigator';
 import HomeScreen from '../screens/Home/HomeScreen.tsx';
 import ProfileScreen from '../screens/Profile/ProfileScreen.tsx';
 import Icon from 'react-native-vector-icons/Ionicons';
+import CatalogScreen from '../screens/Catalog/CatalogScreen.tsx';
 
 const Tab = createBottomTabNavigator();
 
-function TabNavigator() {
+const TabNavigator = () => {
   return (
     <Tab.Navigator initialRouteName="Home" screenOptions={{headerShown: false}}>
       <Tab.Screen
@@ -22,7 +22,7 @@ function TabNavigator() {
       />
       <Tab.Screen
         name="Catalog"
-        component={CatalogStackNavigator}
+        component={CatalogScreen}
         options={{
           tabBarLabel: 'Catalog',
           tabBarIcon: ({color, size}) => (
@@ -42,6 +42,5 @@ function TabNavigator() {
       />
     </Tab.Navigator>
   );
-}
-
+};
 export default TabNavigator;
